[Feat] 첫 data에 로딩 시 로딩화면 구현
